Review 8: Consolidate the skeleton serialization into skeleton.rs
